dawnWeb/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
astropy==4.3.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
beautifulsoup4==4.11.1
bidict==0.22.0
biosppy==0.8.0
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.5
colorlog==6.6.0
cycler==0.11.0
Django==3.2.12
django-pwa==1.0.10
easydev==0.12.0
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
gast==0.3.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.46.3
h5py==2.10.0
hrv-analysis==1.0.4
idna==3.3
importlib-metadata==4.11.4
joblib==1.1.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
line-bot-sdk==2.3.0
Markdown==3.3.7
matplotlib==3.5.2
multidict==6.0.2
nolds==0.5.2
numpy==1.18.5
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pexpect==4.8.0
Pillow==9.1.1
protobuf==3.20.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyerfa==2.0.0.1
pyhrv==0.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.4.1
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
spectrum==0.8.1
sqlparse==0.4.2
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
yarl==1.7.2
zipp==3.8.0
